requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
gunicorn==20.1.0
Flask-Login==0.5.0
sqlalchemy-utils==0.37.8

# Development dependencies
coverage==6.3.2
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0
requests-mock>=1.10.0
bandit>=1.7.5
mypy>=1.5.1
flake8>=6.0.0
black>=24.0.0 